Use library to parse XML into JavaScript Object notation
- it was getting extremely messy and convoluted trying to request data
from Yahoo in json format rather than the default xml format and then
additionally trying to clean up the json data so that it was easy to
work with
- add a dependency to xml-js library that will allow us to better convert
xml strings into js and json objects/strings. there is still a little
cleanup to do in order to have cleaner nested objects and have less
complicated interface definitions but it has already shown to be far easier
to work with than the previous strategy
- add a hook in standings component to check if the selected league has
changed so that we can show/hide the matchups list as necessary. (i.e.
don't persist matchups for a previously selected league. note - matchups
were updating properly if selecting a new league AND THEN also selecting
a new team within that league)
- refactor the matchups component into a league standings component. the
original was starting to get bloated in parsing and managing both league
standings and finding a team's matchups. simplify it so that we split
the 2 into separate components - one for showing standings and then we can
add an onclick effect to pull up a specific team's matchups for the season
- rename ui components to be more concisely named
- slightly adjust/simplify data table component implementation for now
(i was probably just rusty and forgot exactly how it was set up originally...)